priority: Use gnutls_free consistently
The whole point of gnutls_calloc() is to allow an alternative to
malloc() where that alternative takes over all aspects of heap
management; as such, it is never safe to pair bare free() with memory
managed by gnutls.  Not to mention that it looks bad to mix calls to
gnutls_free() and free() to the same variable within the same
function.

Signed-off-by: Eric Blake <eblake@redhat.com>