gnutlsxx: become header-only library
This patch removes the old gnutlsxx library and instead moves all the
definitions of the source file `gnutlsxx.c` to the header file
`gnutlsxx.h`. However, both the C and the C++ library are built. (as
before.)

The user of the C++ interface has two options to choose from:

1. include `gnutlsxx.h` in their application and link against the C
   library. (the default.)
2. include `gnutlsxx.h` in their application, compile with the
   GNUTLS_GNUTLSXX_NO_HEADERONLY macro defined and link against the C++
   library.

Addresses Ref #1381

Signed-off-by: Nikolaos Chatzikonstantinou <nchatz314@gmail.com>