doc: Fix error and improve doc
The description of "gnutls_psk_allocate_client_credentials" incorrectly
states that sc is a pointer to a "gnutls_psk_server_credentials_t". sc
is a pointer to a "gnutls_psk_client_credentials_t".

In the description of "gnutls_credentials_set", mention the type that
the cred parameter should be when using GNUTLS_CRD_PSK.

Signed-off-by: Sahil Siddiq <sahilcdq@proton.me>