leancrypto: support leancrypto for post-quantum algorithms
This adds support for leancrypto as an additional and the preferred
backend for now, until Nettle gains the proper support for PQC
algorithms. There are a few advantages over liboqs, namely:

- It already has required input validations for ML-KEM as in FIPS 203,
  such as Modulus check, which are currently missing in liboqs

- It provides an API to generate ML-KEM/ML-DSA key pairs from a seed,
  which is required to support the seed-only private key format proposed
  in draft-ietf-lamps-dilithium-certificates-05 and later

- No need to avoid undesired OpenSSL dependency; all the symmetric
  algorithms are natively implemented by leancrypto itself

As the supposed use-case of this is to statically link leancrypto with
GnuTLS, this doesn't support loading leancrypto with dlopen.

Signed-off-by: Daiki Ueno <ueno@gnu.org>